Separate tests by initial state
