+++ src/test/java/com/pkb/unit/BasicUnitEventTests.java
@@ -1,6 +1,7 @@
 package com.pkb.unit;
 
 import static com.pkb.unit.DesiredState.UNSET;
+import static com.pkb.unit.State.FAILED;
 import static com.pkb.unit.State.STOPPED;
 import static com.pkb.unit.tracker.ImmutableSystemState.systemState;
 import static com.pkb.unit.tracker.ImmutableUnit.unit;
@@ -19,6 +20,23 @@ public class BasicUnitEventTests extends AbstractUnitTest {
 
         // THEN
         assertLatestState(systemState().addUnits(
-                unit("unit1").withState(STOPPED).withDesiredState(UNSET)).build());
+                unit("unit1").withState(STOPPED).withDesiredState(UNSET))
+                .build());
+    }
+
+    @Test
+    public void whenAdvertiseFailedOnStoppedThenFailed() throws Exception {
+        // GIVEN
+        setupComputationAndIOTestScheduler();
+
+        // WHEN
+        FakeUnit fakeUnit = new FakeUnit("unit1", bus);
+        fakeUnit.failed();
+        testScheduler.triggerActions();
+
+        // THEN
+        assertLatestState(systemState().addUnits(
+                unit("unit1").withDesiredState(UNSET).withState(FAILED))
+                .build());
     }
 }
