+++ src/test/java/com/pkb/unit/FailedTransitionTests.java
@@ -1,0 +1,66 @@
+package com.pkb.unit;
+
+import static com.pkb.unit.Command.START;
+import static com.pkb.unit.Command.STOP;
+import static com.pkb.unit.DesiredState.UNSET;
+import static com.pkb.unit.State.FAILED;
+import static com.pkb.unit.tracker.ImmutableSystemState.systemState;
+import static com.pkb.unit.tracker.ImmutableUnit.unit;
+
+import org.junit.Test;
+
+public class FailedTransitionTests extends AbstractUnitTest {
+
+    // I/7
+    @Test
+    public void whenStartFailedUnsetAndFailStartThenFailed() throws Exception {
+        // GIVEN
+        setupComputationAndIOTestScheduler();
+
+        // WHEN
+        FakeUnit fakeUnit = new FakeUnit("unit1", bus);
+        bus.sink().accept(command("unit1", START));
+        fakeUnit.failStart();
+        testScheduler.triggerActions();
+
+        // THEN
+        assertLatestState(systemState().addUnits(
+                unit("unit1").withDesiredState(UNSET).withState(FAILED))
+                .build());
+
+        // WHEN
+        testScheduler.advanceTimeBy(FakeUnit.RETRY_PERIOD + 1, FakeUnit.RETRY_PERIOD_UNIT);
+
+        // THEN
+        assertLatestState(systemState().addUnits(
+                unit("unit1").withDesiredState(UNSET).withState(FAILED))
+                .build());
+    }
+
+    // I/11
+    @Test
+    public void whenStopFailedUnsetThenFailed() throws Exception {
+        // GIVEN
+        setupComputationAndIOTestScheduler();
+
+        // WHEN
+        FakeUnit fakeUnit = new FakeUnit("unit1", bus);
+        bus.sink().accept(command("unit1", START));
+        fakeUnit.failStart();
+        testScheduler.triggerActions();
+
+        // THEN
+        assertLatestState(systemState().addUnits(
+                unit("unit1").withDesiredState(UNSET).withState(FAILED))
+                .build());
+
+        // WHEN
+        bus.sink().accept(command("unit1", STOP));
+        testScheduler.triggerActions();
+
+        // THEN
+        assertLatestState(systemState().addUnits(
+                unit("unit1").withDesiredState(UNSET).withState(FAILED))
+                .build());
+    }
+}
