+++ src/test/java/com/pkb/unit/StartedTransitionTests.java
@@ -1,0 +1,160 @@
+package com.pkb.unit;
+
+import static com.pkb.unit.Command.DISABLE;
+import static com.pkb.unit.Command.ENABLE;
+import static com.pkb.unit.Command.START;
+import static com.pkb.unit.Command.STOP;
+import static com.pkb.unit.DesiredState.DISABLED;
+import static com.pkb.unit.DesiredState.ENABLED;
+import static com.pkb.unit.DesiredState.UNSET;
+import static com.pkb.unit.State.FAILED;
+import static com.pkb.unit.State.STARTED;
+import static com.pkb.unit.State.STOPPED;
+import static com.pkb.unit.State.STOPPING;
+import static com.pkb.unit.tracker.ImmutableSystemState.systemState;
+import static com.pkb.unit.tracker.ImmutableUnit.unit;
+
+import org.junit.Test;
+
+public class StartedTransitionTests extends AbstractUnitTest {
+
+    // I/3
+    @Test
+    public void whenStartStartedUnsetThenStarted() throws Exception {
+        // GIVEN
+        setupComputationAndIOTestScheduler();
+
+        // WHEN
+        FakeUnit unit1 = new FakeUnit("unit1", bus);
+        bus.sink().accept(command("unit1", START));
+        unit1.completeStart();
+        testScheduler.triggerActions();
+
+        // THEN
+        assertLatestState(systemState().addUnits(
+                unit("unit1").withState(STARTED).withDesiredState(UNSET))
+                .build());
+
+        // WHEN
+        bus.sink().accept(command("unit1", START));
+        testScheduler.triggerActions();
+
+        // THEN
+        assertLatestState(systemState().addUnits(
+                unit("unit1").withState(STARTED).withDesiredState(UNSET))
+                .build());
+    }
+
+    // I/4
+    @Test
+    public void whenDisableStartedAndCompleteStopStoppingThenStopped() throws Exception {
+        // GIVEN
+        setupComputationTestScheduler();
+        setupIOTestScheduler();
+
+        // WHEN
+        FakeUnit fakeUnit = new FakeUnit("unit1", bus);
+        bus.sink().accept(command("unit1", ENABLE));
+        fakeUnit.completeStart();
+        testComputationScheduler.triggerActions();
+        testIOScheduler.triggerActions();
+
+        // THEN
+        assertLatestState(systemState().addUnits(
+                unit("unit1").withDesiredState(ENABLED).withState(STARTED))
+                .build());
+
+        // WHEN
+        bus.sink().accept(command("unit1", DISABLE));
+        fakeUnit.completeStop();
+        testComputationScheduler.triggerActions();
+        testIOScheduler.triggerActions();
+
+        // THEN
+        assertLatestState(systemState().addUnits(
+                unit("unit1").withDesiredState(DISABLED).withState(STOPPED))
+                .build());
+    }
+
+    // I/7
+    @Test
+    public void whenStopStartedEnabledAndFaileStopAndStartRetryThenStarted() throws Exception {
+        // GIVEN
+        setupComputationTestScheduler();
+        setupIOTestScheduler();
+
+        // WHEN
+        FakeUnit fakeUnit = new FakeUnit("unit1", bus);
+        bus.sink().accept(command("unit1", ENABLE));
+        fakeUnit.completeStart();
+        testComputationScheduler.triggerActions();
+        testIOScheduler.triggerActions();
+
+        // THEN
+        assertLatestState(systemState().addUnits(
+                unit("unit1").withDesiredState(ENABLED).withState(STARTED))
+                .build());
+
+        // WHEN
+        bus.sink().accept(command("unit1", STOP));
+        fakeUnit.failStop();
+        testComputationScheduler.triggerActions();
+        testIOScheduler.triggerActions();
+
+        // THEN
+        assertLatestState(systemState().addUnits(
+                unit("unit1").withDesiredState(ENABLED).withState(FAILED))
+                .build());
+
+        // WHEN
+        fakeUnit.completeStart();
+        testComputationScheduler.advanceTimeBy(FakeUnit.RETRY_PERIOD + 1, FakeUnit.RETRY_PERIOD_UNIT);
+        testIOScheduler.triggerActions();
+
+        // THEN
+        assertLatestState(systemState().addUnits(
+                unit("unit1").withDesiredState(ENABLED).withState(STARTED))
+                .build());
+    }
+
+    // II/12
+    @Test
+    public void whenStopStartedUnsetAndStartStoppingThenStopped() throws Exception {
+        // GIVEN
+        setupComputationTestScheduler();
+        setupIOTestScheduler();
+
+        // WHEN
+        FakeUnit unit1 = new FakeUnit("unit1", bus);
+        bus.sink().accept(command("unit1", START));
+        unit1.completeStart();
+        testComputationScheduler.triggerActions();
+        testIOScheduler.triggerActions();
+
+        // THEN
+        assertLatestState(systemState().addUnits(
+                unit("unit1").withState(STARTED).withDesiredState(UNSET))
+                .build());
+
+        // WHEN
+        bus.sink().accept(command("unit1", STOP));
+        testComputationScheduler.triggerActions();
+
+        // THEN
+        assertLatestState(systemState().addUnits(
+                unit("unit1").withState(STOPPING).withDesiredState(UNSET))
+                .build());
+
+        // WHEN
+        bus.sink().accept(command("unit1", START));
+        unit1.completeStart();
+        unit1.completeStop();
+        testComputationScheduler.triggerActions();
+        testIOScheduler.triggerActions();
+
+        // THEN
+        assertLatestState(systemState().addUnits(
+                unit("unit1").withState(STOPPED).withDesiredState(UNSET))
+                .build());
+    }
+}
